refactor: properly modularize content script for testability
- Extract testable logic into content-logic.ts module
- Update content.ts to import from content-logic
- Fix content.test.ts to import from content-logic instead of reimplementing
- All tests, typecheck, and lint passing locally
